finished improving unit tests +  modifying methods for general.py.
Former-commit-id: 4efa938764186ff4ed278595e0905238b89f24ee

+++ ivy/numpy/core/general.py
@@ -13,7 +13,7 @@ DTYPE_DICT = {_np.dtype('bool'): 'bool',
               _np.dtype('int16'): 'int16',
               _np.dtype('int32'): 'int32',
               _np.dtype('int64'): 'int64',
-              _np.dtype('float16'): 'float32',
+              _np.dtype('float16'): 'float16',
               _np.dtype('float32'): 'float32',
               _np.dtype('float64'): 'float64'}
